fixed some typos, added xml comments, fixed some class names
