+++ Navferty.Common/Controls/ErrorDialog.resx
