+++ Navferty.Common/Extensions/ErrorsExtensions.cs
@@ -17,6 +17,11 @@ namespace Navferty.Common
 	[DebuggerStepThrough]
 	public static class ErrorsExtensions
 	{
+		/// <summary>Displays error message box to user, with ability to send feedback, and wites error to the Log</summary>
+		/// <param name="errorTitle">Some description about error to display to the user</param>
+		/// <param name="logger"></param>
+		/// <param name="loggerTitle">Some description about error to write to Log. If null - use errorTitle</param>
+		/// <param name="allowErrorReporting">Allow user to send bug report</param>
 		public static void ShowErrorUI(
 			this Exception ex,
 			string? errorTitle = null,
@@ -24,13 +29,11 @@ namespace Navferty.Common
 			string? loggerTitle = null,
 			bool allowErrorReporting = true)
 		{
-
-			//errorTitle ??= Application.ProductName;
 			loggerTitle ??= errorTitle ?? "[Error description empty]";
 			logger ??= NLog.LogManager.GetCurrentClassLogger();
 			logger?.Error(ex, loggerTitle);
 
-			using var ef = new ErrorForm(ex, errorTitle, allowErrorReporting);
+			using var ef = new ErrorDialog(ex, errorTitle, allowErrorReporting);
 			ef.ShowDialog();
 		}
 	}
