+++ Navferty.Common/WinAPI/WinAPI.Core.cs
