+++ Navferty.Common/WinAPI/WinAPI.GDI.cs
